feat: enabled drop target for FolderItem
