+++ package.json
@@ -89,7 +89,7 @@
         "xvfb-maybe": "^0.2.1"
     },
     "dependencies": {
-        "@powerhousedao/design-system": "1.0.0-alpha.75",
+        "@powerhousedao/design-system": "1.0.0-alpha.76",
         "did-key-creator": "^1.2.0",
         "document-drive": "1.0.0-alpha.17",
         "document-model": "1.0.32",
